Added functions to split an overall sample into multiple bitstrings.
The idea is to get a sample separated by registers. Even though it is
already possible to use the qubits flag in the to_job() function, in
this way we lose the ability to retrieve the original amplitude.
Using this function we can split the resulting bitstring afterwards,
separating the bitstring into its qregs.

In addition, we can use to name_to_reg dictionary if we have previously
associated a name to a specific qreg.