gje: fix gates ordering for non-Prange ISDs
In improvement 5, in each last iteration of phase 1, we postpone the pivot
pseudo swap until the last operation. However, this operation should be done as
well before the latest k columns.

In improvement 6, we do the same in the last iteration of phase 2, but only for
the second-to-last iteration row. In this way, we can anticipate last iteration
gates. Same consideration on the operations on the k columns as before.